requirements.txt
django
djangorestframework
drf-nested-routers
aiogram